# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "X-Agent API",
    "description": "Complete REST API for X-Agent Core - An autonomous agent framework with workflow orchestration, memory management, and tool integration",
    "version": "1.0.0",
    "contact": {
      "name": "X-Agent Team",
      "url": "https://github.com/x-agent"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://api.x-agent.io",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/memory": {
      "post": {
        "tags": ["Memory"],
        "summary": "Store memory",
        "operationId": "storeMemory",
        "security": [
          {"ApiKeyAuth": []},
          {"BearerAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/MemoryStoreRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Memory stored",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/MemoryStoreResponse"}
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/search": {
      "post": {
        "tags": ["Memory"],
        "summary": "Search memory",
        "operationId": "searchMemory",
        "security": [
          {"ApiKeyAuth": []},
          {"BearerAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/MemorySearchRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/MemorySearchResponse"}
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/export": {
      "get": {
        "tags": ["Memory"],
        "summary": "Export memory",
        "operationId": "exportMemory",
        "security": [
          {"ApiKeyAuth": []},
          {"BearerAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Memory export",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/MemoryExportResponse"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MemoryStoreRequest": {
        "type": "object",
        "properties": {
          "content": {"type": "string", "minLength": 1, "maxLength": 20000},
          "layer": {"type": "integer", "minimum": 1, "maximum": 10, "default": 3},
          "importance": {"type": "number", "minimum": 0, "maximum": 1, "default": 0.5},
          "tags": {"type": "array", "items": {"type": "string"}},
          "metadata": {"type": "object"},
          "session_id": {"type": "string"},
          "scope": {"type": "string", "enum": ["user", "session", "global"]}
        },
        "required": ["content"]
      },
      "MemoryStoreResponse": {
        "type": "object",
        "properties": {
          "id": {"type": "string"}
        }
      },
      "MemorySearchRequest": {
        "type": "object",
        "properties": {
          "query": {"type": "string"},
          "layers": {"type": "array", "items": {"type": "integer"}},
          "top_k": {"type": "integer", "minimum": 1, "maximum": 50, "default": 5},
          "include_scores": {"type": "boolean", "default": false},
          "scope": {"type": "string"}
        }
      },
      "MemorySearchResponse": {
        "type": "object",
        "properties": {
          "items": {"type": "array"},
          "hits": {"type": "array"}
        }
      },
      "MemoryExportResponse": {
        "type": "object",
        "properties": {
          "bundle": {"type": "object"}
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}